Split josh-changes lib.rs into modules by responsibility
Move the 2288-line lib.rs into focused modules: refs (ref-name
grammars), change (Change model and discovery), store (changes-ref
tree I/O), comments, votes, revisions, union (cross-scope merge
helpers), and forges/{gerrit,github} for forge-specific publishing.
The public API is unchanged via flat re-exports at the crate root.

Change: split-josh-changes-modules
